requirements.txt
tensorflow
streamlit
numpy
pillow
plotly
<<<<<<< HEAD
=======
huggingface_hub
>>>>>>> d77b7f0 (upload req and utils)
